Refactor firewall backends into dedicated resources
