+++ .github/workflows/ci.yml
@@ -7,7 +7,7 @@ name: ci
 
 jobs:
   lint-unit:
-    uses: sous-chefs/.github/.github/workflows/lint-unit.yml@6.0.0
+    uses: sous-chefs/.github/.github/workflows/lint-unit.yml@8.0.0
     permissions:
       checks: write
       pull-requests: write
@@ -192,7 +192,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v6
       - name: Install Cinc Workstation
-        uses: sous-chefs/.github/.github/actions/install-workstation@6.0.0
+        uses: sous-chefs/.github/.github/actions/install-workstation@8.0.0
       - name: Dokken
         uses: actionshub/test-kitchen@3.0.0
         env:
