+++ documentation/resource_nftables.md
@@ -10,7 +10,7 @@
 ## Properties
 
 | Name           | Type               | Default  | Description                                                      |
-| -------------- | ------------------ | -------- | ---------------------------------------------------------------- |
+|----------------|--------------------|----------|------------------------------------------------------------------|
 | rules          | Hash               | {}       | Rules are accumulated in here. Key is a rule, value its priority |
 | input_policy   | ['drop', 'accept'] | 'accept' | Policy for the input chain                                       |
 | output_policy  | ['drop', 'accept'] | 'accept' | Policy for the output chain                                      |
