+++ resources/nftables_rule.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 unified_mode true
 
 require 'ipaddr'
@@ -90,24 +92,21 @@ property :log_group,
 property :raw,
          String
 
-# do you want this rule to notify the firewall to recalculate
-# (and potentially reapply) the firewall_rule(s) it finds?
+# do you want this rule to notify nftables to recalculate
+# (and potentially reapply) the nftables_rule(s) it finds?
 property :notify_firewall,
          [true, false],
          default: true
 
 action :create do
   return if return_early?(new_resource)
-  fwr = build_firewall_rule(new_resource)
 
-  with_run_context :root do
-    edit_resource!('nftables', new_resource.firewall_name) do |fw_rule|
-      r = rules.dup || {}
-      r.merge!({
-                 fwr => fw_rule.position,
-               })
-      rules(r)
-      delayed_action :rebuild
-    end
+  Chef.run_context.resource_collection.find(nftables: new_resource.firewall_name)
+  build_firewall_rule(new_resource)
+
+  ruby_block "queue nftables rebuild #{new_resource.firewall_name} #{new_resource.name}" do
+    block {}
+    action :run
+    notifies :rebuild, "nftables[#{new_resource.firewall_name}]", :delayed
   end
 end
