+++ spec/unit/resources/nftables_rule_spec.rb
@@ -15,4 +15,5 @@ describe 'nftables_rule' do
   end
 
   it { is_expected.to create_nftables_rule('ssh') }
+  it { is_expected.to run_ruby_block('queue nftables rebuild default ssh') }
 end
